requirements.txt
discord.py==2.1.1
Flask==3.0.2